+++ .eslintrc.json
@@ -14,6 +14,7 @@
     },
     "plugins": ["react", "prettier"],
     "rules": {
-        "react/prop-types": "off"
+        "react/prop-types": "off",
+        "no-unused-vars": "off"
     }
 }
