+++ src/components/Movies.js
@@ -1,21 +1,34 @@
 import React, { useState, useEffect } from 'react'
 import MovieCard from './MovieCard'
+import Pagination from './Pagination'
 import axios from 'axios'
 import { TailSpin } from 'react-loader-spinner'
 
 const Movies = () => {
     const [movies, setMovies] = useState([])
 
+    const [page, setPage] = useState(1)
+
+    const goAhead = () => {
+        setPage(page + 1)
+    }
+
+    const goBack = () => {
+        if (page > 1) {
+            setPage(page - 1)
+        }
+    }
+
     useEffect(() => {
         axios
             .get(
-                'https://api.themoviedb.org/3/trending/movie/week?api_key=c51ead54fa195e58d0023669954222a3&page=1'
+                `https://api.themoviedb.org/3/trending/movie/week?api_key=c51ead54fa195e58d0023669954222a3&page=${page}`
             )
             .then((res) => {
                 console.table(res.data.results)
                 setMovies(res.data.results)
             })
-    })
+    }, [page])
     return (
         <>
             <div className="mb-8">
@@ -39,6 +52,7 @@ const Movies = () => {
                     </div>
                 )}
             </div>
+            <Pagination page={page} goAhead={goAhead} goBack={goBack} />
         </>
     )
 }
