+++ src/pages/Home.js
@@ -1,14 +1,12 @@
 import React from 'react'
 import Banner from '../components/Banner'
 import Movies from '../components/Movies'
-import Pagination from '../components/Pagination'
 
 const Home = () => {
     return (
         <>
             <Banner />
             <Movies />
-            <Pagination />
         </>
     )
 }
